reduce the redundant code
